RONDB-671: Improved interpreter
When performing arithmetic interpreter operations we will always
work on two registers. Previously it was enough for one of them
to be not NULL, now it is checked that both registers are not
NULL before performing the operation. If any of them are NULL
the interpreter will fail with a ZREGISTER_INIT_ERROR.